Move SpanRectangle into its own file
