+++ static/app/components/events/interfaces/spans/spanSiblingGroupBar.tsx
@@ -7,7 +7,6 @@ import {
   RowCell,
   RowCellContainer,
 } from 'sentry/components/performance/waterfall/row';
-import {DurationPill, RowRectangle} from 'sentry/components/performance/waterfall/rowBar';
 import {
   DividerContainer,
   DividerLine,
@@ -25,26 +24,20 @@ import {
   TreeToggle,
   TreeToggleContainer,
 } from 'sentry/components/performance/waterfall/treeConnector';
-import {
-  getDurationDisplay,
-  getHumanDuration,
-  toPercent,
-} from 'sentry/components/performance/waterfall/utils';
+import {toPercent} from 'sentry/components/performance/waterfall/utils';
 import {t} from 'sentry/locale';
 import {EventTransaction} from 'sentry/types/event';
 import {defined} from 'sentry/utils';
-import theme from 'sentry/utils/theme';
 
 import * as DividerHandlerManager from './dividerHandlerManager';
 import * as ScrollbarManager from './scrollbarManager';
 import SpanBarCursorGuide from './spanBarCursorGuide';
+import SpanRectangle from './spanRectangle';
 import {MeasurementMarker} from './styles';
 import {EnhancedSpan, ProcessedSpanType, SpanType, TreeDepthType} from './types';
 import {
   getMeasurementBounds,
   getMeasurements,
-  getSpanGroupBounds,
-  getSpanGroupTimestamps,
   getSpanOperation,
   isOrphanSpan,
   isOrphanTreeDepth,
@@ -335,85 +328,4 @@ export default function SpanSiblingGroupBar(props: Props) {
       )}
     </ScrollbarManager.Consumer>
   );
-
-  function SpanRectangle({
-    spanGrouping,
-    index,
-    generateBounds,
-  }: {
-    generateBounds: Props['generateBounds'];
-    spanGrouping: EnhancedSpan[];
-    index?: number;
-  }) {
-    // If an index is provided, we treat this as an individual block for a single span in the grouping
-    if (index !== undefined) {
-      const grouping = [spanGrouping[index]];
-      const bounds = getSpanGroupBounds(grouping, generateBounds);
-
-      // If this is not the last span in the grouping, the duration does not need to be rendered
-      if (index !== spanGrouping.length - 1) {
-        return (
-          <RowRectangle
-            spanBarHatch={false}
-            style={{
-              backgroundColor: theme.blue300,
-              left: `min(${toPercent(bounds.left || 0)}, calc(100% - 1px))`,
-              width: toPercent(bounds.width || 0),
-            }}
-          />
-        );
-      }
-
-      const {startTimestamp, endTimestamp} = getSpanGroupTimestamps(spanGrouping);
-      const duration = Math.abs(endTimestamp - startTimestamp);
-      const durationDisplay = getDurationDisplay(bounds);
-      const durationString = getHumanDuration(duration);
-
-      return (
-        <RowRectangle
-          spanBarHatch={false}
-          style={{
-            backgroundColor: theme.blue300,
-            left: `min(${toPercent(bounds.left || 0)}, calc(100% - 1px))`,
-            width: toPercent(bounds.width || 0),
-          }}
-        >
-          {index === spanGrouping.length - 1 && (
-            <DurationPill
-              durationDisplay={durationDisplay}
-              showDetail={false}
-              spanBarHatch={false}
-            >
-              {durationString}
-            </DurationPill>
-          )}
-        </RowRectangle>
-      );
-    }
-    const bounds = getSpanGroupBounds(spanGrouping, generateBounds);
-    const durationDisplay = getDurationDisplay(bounds);
-    const {startTimestamp, endTimestamp} = getSpanGroupTimestamps(spanGrouping);
-    const duration = Math.abs(endTimestamp - startTimestamp);
-    const durationString = getHumanDuration(duration);
-    return (
-      <RowRectangle
-        spanBarHatch={false}
-        style={{
-          backgroundColor: theme.blue300,
-          left: `min(${toPercent(bounds.left || 0)}, calc(100% - 1px))`,
-          width: toPercent(bounds.width || 0),
-        }}
-      >
-        {
-          <DurationPill
-            durationDisplay={durationDisplay}
-            showDetail={false}
-            spanBarHatch={false}
-          >
-            {durationString}
-          </DurationPill>
-        }
-      </RowRectangle>
-    );
-  }
 }
